requirements.txt
mutagen
python-dotenv
schedule